Changed the index.html to now be global.
