--- backend/quiz/templates/quiz/index.html
@@ -1,32 +1,0 @@
-{% load static %}
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-	<link href="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-icons/1.8.1/font/bootstrap-icons.min.css" rel="stylesheet">
-
-	<link href="{% static 'quiz/css/room_list.css' %}" rel="stylesheet">
-	<link href="{% static 'quiz/css/room_display.css' %}" rel="stylesheet">
-	<link href="{% static 'quiz/css/quiz_game.css' %}" rel="stylesheet">
-
-	<meta name="csrf-token" content="{{ csrf_token }}">
-	<title>Quiz Home</title>
-</head>
-
-<body>
-	{% include 'navbar.html' %}
-
-	<div class="container mt-5" id="quiz-app-content"></div>
-
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-	<!-- <script src="{% static 'quiz/js/create_room.js' %}"></script>
-	<script src="{% static 'quiz/js/room_list.js' %}"></script> -->
-	<script type="module" src="{% static 'quiz/js/router.js' %}"></script>
-	<!-- <script type="module" src="{% static 'quiz/js/room_list.js' %}"></script> -->
-
-</body>
-
-</html>
