+++ backend/templates/index.html
@@ -39,7 +39,7 @@
 						<a class="nav-item nav-link" href="">Home</a>
 						<a class="nav-item nav-link" href="/chat">Chat</a>
 						<a class="nav-item nav-link" href="/about">About</a>
-						<a class="nav-item nav-link" href="/quiz">Quiz App</a>
+						<a class="nav-item nav-link" href="/quiz/">Quiz App</a>
 						<a class="nav-item nav-link" href="{% url 'game:recent_games' %}">Game</a>
 					</div>
 					<!-- Navbar Right Side -->
@@ -64,7 +64,7 @@
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
 
 
-	<script type="module" src="{% static '/js/router.js' %}"></script>
+	<script type="module" src="{% static 'js/router.js' %}"></script>
 
 </body>
