+++ backend/transcendence/urls.py
@@ -15,13 +15,14 @@ Including another URLconf
 	2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path, include
+from django.urls import path, include, re_path
 from django.conf import settings
 from debug_toolbar.toolbar import debug_toolbar_urls
 from . import views
 
 urlpatterns = [
 	path('', views.index, name='index'),
+	path('quiz/', views.index, name='quiz'),
 	path('users/', include("users.urls")),
 	path('game/', include("game.urls")),
 	path('admin/', admin.site.urls),
@@ -29,6 +30,7 @@ urlpatterns = [
 	path('chat/', include('chat.urls')),
 	path('quiz/', include('quiz.urls')),
     # path('__debug__/', include(debug_toolbar.urls)),
+	re_path(r'^.*$', views.index, name='index'),
 ]
